+++ apps/client/src/app/components/access-table/access-table.module.ts
@@ -1,5 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { MatMenuModule } from '@angular/material/menu';
 import { MatTableModule } from '@angular/material/table';
 
 import { AccessTableComponent } from './access-table.component';
@@ -7,7 +9,7 @@ import { AccessTableComponent } from './access-table.component';
 @NgModule({
   declarations: [AccessTableComponent],
   exports: [AccessTableComponent],
-  imports: [CommonModule, MatTableModule],
+  imports: [CommonModule, MatButtonModule, MatMenuModule, MatTableModule],
   providers: [],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })