+++ apps/client/src/app/pages/account/create-or-update-access-dialog/interfaces/interfaces.ts
@@ -1,0 +1,5 @@
+import { Access } from '@ghostfolio/common/interfaces';
+
+export interface CreateOrUpdateAccessDialogParams {
+  access: Access;
+}